requirements.txt
Appium-Python-Client==0.48
selenium==3.141.0
urllib3==1.25.5
selenium-page-factory==1.2
self==2019.4.13
